+++ es/rememo.js
@@ -1,6 +1,8 @@
 'use strict';
 
 import memoize from 'moize';
+import { getMultiParamKey } from 'moize/lib/utils';
+import Cache from 'moize/lib/Cache';
 import shallowEqual from 'shallow-equal/arrays';
 
 /**
@@ -11,6 +13,30 @@ import shallowEqual from 'shallow-equal/arrays';
  */
 function identity( value ) {
 	return value;
+}
+
+/**
+ * Creates a custom serializer for memoization.
+ *
+ * @param  {moize.Cache} cache Moize cache instance
+ * @return {Function}          Memoization serializer
+ */
+function createSerializer( cache ) {
+	/**
+	 * Serializes arguments for generating a cache key. This behaves more-or-
+	 * less identical to moize's default `getCacheKey` behavior, with the
+	 * exception that it ignores the first argument in generating the key.
+	 *
+	 * @param  {Array} args Memoized function arguments as array
+	 * @return {*}          Cache key
+	 */
+	return function( args ) {
+		if ( args.length > 2 ) {
+			return getMultiParamKey( cache, args.slice( 1 ) );
+		}
+
+		return args[ 1 ];
+	};
 }
 
 /**
@@ -28,8 +54,17 @@ function identity( value ) {
  * @return {*}                      Selector return value
  */
 export default function( selector, getDependants ) {
-	var memoizedSelector = memoize( selector ),
-		lastDependants;
+	var cache, memoizedSelector, lastDependants;
+
+	// We need to maintain our own cache in order to recreate the default cache
+	// serialization behavior using moize's `getMultiParamKey` utility
+	cache = new Cache();
+
+	memoizedSelector = memoize( selector, {
+		cache: cache,
+		serialize: true,
+		serializer: createSerializer( cache )
+	} );
 
 	if ( ! getDependants ) {
 		getDependants = identity;