+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#### v1.1.0 (2017-06-08)
+
+- Improved: Object target is ignored in generating memoized function cache key.
+This can resolve issues where cache would be discarded if dependant references
+were the same but the target object reference changed.
+
 #### v1.0.2 (2017-05-29)
 
 - Fix: Include dist in npm package (for unpkg availability)